Реализация сортировки по Дню начала. Также немного изменила фильтры - view и разделила utils, так как слишком большой стал
