+++ src/presenter/event-presenter.js
@@ -1,7 +1,7 @@
 import EventItemView from '../view/event-view.js';
 import FormEditView from '../view/form-edit-view.js';
 import { render, replace, remove } from '../framework/render.js';
-import { isEscapeKey } from '../utils.js';
+import { isEscapeKey } from '../utils/utils.js';
 import { Mode } from '../const.js';
 
 export default class EventPresenter {
