+++ src/view/filter-view.js
@@ -1,27 +1,20 @@
 import AbstractView from '../framework/view/abstract-view.js';
+import { Filters } from '../const.js';
+import { capitalizeWords } from '../utils/utils.js';
 
-function createEventFilterTemplate() {
+function createFilterTemplate(filterName) {
+  return (
+    `<div class="trip-filters__filter">
+      <input id="filter-${filterName}" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="${filterName}">
+      <label class="trip-filters__filter-label" for="filter-${filterName}">${capitalizeWords(filterName)}</label>
+    </div>`
+  );
+}
+
+function createEventFiltersTemplate() {
   return (
     `<form class="trip-filters" action="#" method="get">
-      <div class="trip-filters__filter">
-        <input id="filter-everything" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="everything">
-        <label class="trip-filters__filter-label" for="filter-everything">Everything</label>
-      </div>
-
-      <div class="trip-filters__filter">
-        <input id="filter-future" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="future">
-        <label class="trip-filters__filter-label" for="filter-future">Future</label>
-      </div>
-
-      <div class="trip-filters__filter">
-         <input id="filter-present" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="present">
-         <label class="trip-filters__filter-label" for="filter-present">Present</label>
-      </div>
-
-      <div class="trip-filters__filter">
-        <input id="filter-past" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="past" checked>
-        <label class="trip-filters__filter-label" for="filter-past">Past</label>
-      </div>
+      ${Object.values(Filters).map((filter) => createFilterTemplate(filter)).join('')}
 
       <button class="visually-hidden" type="submit">Accept filter</button>
     </form>`
@@ -30,6 +23,6 @@ function createEventFilterTemplate() {
 
 export default class EventFilterView extends AbstractView {
   get template() {
-    return createEventFilterTemplate();
+    return createEventFiltersTemplate();
   }
 }
